requirements.txt
fastapi
uvicorn
openai
python-dotenv
crawl4ai
